### 2021-01-30 #2997 @whedon
[PRE REVIEW]: seaborn: statistical data visualization
**Submitting author:** @mwaskom (<a href="http://orcid.org/0000-0002-9817-6869">Michael Waskom</a>)
**Repository:** <a href="https://github.com/mwaskom/seaborn" target ="_blank">https://github.com/mwaskom/seaborn</a>
**Version:** v0.11.1
