### 2021-01-30 #2997 @danielskatz
@mwaskom - there is at least one reference that needs to be fixed, and the missing DOIs also might be correct. Please feel free to fix them, then use the @whedon commands I used to regenerate the PDF and recheck the references.
