+++ paper.bib
@@ -7,6 +7,7 @@ year = {2005},
 volume = {60},
 number = {2},
 pages = {170--180},
+doi = {10.1037/0003-066X.60.2.170},
 }
 
 @Article{         Harris:2020,
@@ -54,22 +55,15 @@ pages = {170--180},
   pages     = { 51 - 56 },
   year      = { 2010 },
   editor    = { St\'efan van der Walt and Jarrod Millman }
+  doi = {10.25080/Majora-92bf1922-00a},
 }
 
 @book{Tukey:1977,
-  added-at = {2009-10-28T04:42:52.000+0100},
   author = {Tukey, John W.},
-  biburl = {https://www.bibsonomy.org/bibtex/2e0c6e346a4594eb8f0bc71a6e65366fb/jwbowers},
-  citeulike-article-id = {107137},
-  date-added = {2007-09-03 22:45:16 -0500},
-  date-modified = {2007-09-03 22:45:16 -0500},
-  interhash = {69fff218408c5eda560e267ff0099dec},
-  intrahash = {e0c6e346a4594eb8f0bc71a6e65366fb},
-  keywords = {eda statistics},
   publisher = {Addison-Wesley},
-  timestamp = {2009-10-28T04:43:20.000+0100},
   title = {Exploratory Data Analysis},
-  year = 1977
+  isbn = {978-0201076165},
+  year = 1977,
 }
 
 @article{Waskom:2017,
@@ -78,7 +72,7 @@ title = {{Distributed representation of context by intrinsic subnetworks in pref
 journal = {Proceedings of the National Academy of Sciences},
 year = {2017},
 pages = {2030--2035},
-doi = {https://doi.org/10.1073/pnas.1615269114},
+doi = {10.1073/pnas.1615269114},
 month = feb
 }
 
@@ -91,7 +85,6 @@ month = feb
    number = {10},
    year = {2014},
    keywords = {},
-   abstract = {A huge amount of effort is spent cleaning data to get it ready for analysis, but there has been little research on how to make data cleaning as easy and effective as possible. This paper tackles a small, but important, component of data cleaning: data tidying. Tidy datasets are easy to manipulate, model and visualize, and have a specific structure: each variable is a column, each observation is a row, and each type of observational unit is a table. This framework makes it easy to tidy messy datasets because only a small set of tools are needed to deal with a wide range of un-tidy datasets. This structure also makes it easier to develop tidy tools for data analysis, tools that both input and output tidy datasets. The advantages of a consistent data structure and matching tools are demonstrated with a case study free from mundane data manipulation chores.},
    issn = {1548-7660},
    pages = {1--23},
    doi = {10.18637/jss.v059.i10},
