+++ paper.bib
@@ -54,7 +54,7 @@ doi = {10.1037/0003-066X.60.2.170},
   booktitle = { Proceedings of the 9th Python in Science Conference },
   pages     = { 51 - 56 },
   year      = { 2010 },
-  editor    = { St\'efan van der Walt and Jarrod Millman }
+  editor    = { St\'efan van der Walt and Jarrod Millman },
   doi = {10.25080/Majora-92bf1922-00a},
 }
 
