### 2021-01-30 #2997 @mwaskom
@danielskatz  broken references are now fixed ... Exploratory Data Analysis is a book from the 1970s so I don't think it has a DOI?
