+++ seaborn/distributions.py
@@ -30,6 +30,7 @@ from .utils import (
     _normalize_kwargs,
     _check_argument,
     _assign_default_kwargs,
+    _default_color,
 )
 from .palettes import color_palette
 from .external import husl
@@ -141,11 +142,18 @@ class _DistributionPlotter(VectorPlotter):
         labels = []
         for level in self._hue_map.levels:
             color = self._hue_map(level)
-            handles.append(artist(
-                **self._artist_kws(
-                    artist_kws, fill, element, multiple, color, alpha
-                )
-            ))
+
+            kws = self._artist_kws(
+                artist_kws, fill, element, multiple, color, alpha
+            )
+
+            # color gets added to the kws to workaround an issue with barplot's color
+            # cycle integration but it causes problems in this context where we are
+            # setting artist properties directly, so pop it off here
+            if "facecolor" in kws:
+                kws.pop("color", None)
+
+            handles.append(artist(**kws))
             labels.append(level)
 
         if isinstance(ax_obj, mpl.axes.Axes):
@@ -163,7 +171,14 @@ class _DistributionPlotter(VectorPlotter):
         """Handle differences between artists in filled/unfilled plots."""
         kws = kws.copy()
         if fill:
+            kws = _normalize_kwargs(kws, mpl.collections.PolyCollection)
             kws.setdefault("facecolor", to_rgba(color, alpha))
+
+            if element == "bars":
+                # Make bar() interface with property cycle correctly
+                # https://github.com/matplotlib/matplotlib/issues/19385
+                kws["color"] = "none"
+
             if multiple in ["stack", "fill"] or element == "bars":
                 kws.setdefault("edgecolor", mpl.rcParams["patch.edgecolor"])
             else:
@@ -398,7 +413,7 @@ class _DistributionPlotter(VectorPlotter):
         else:
             common_norm = False
 
-        # Turn multiple off if no hue or if hue exists but is redundent with faceting
+        # Turn multiple off if no hue or if hue exists but is redundant with faceting
         facet_vars = [self.variables.get(var, None) for var in ["row", "col"]]
         if "hue" not in self.variables:
             multiple = None
@@ -486,34 +501,6 @@ class _DistributionPlotter(VectorPlotter):
 
         # Note: default linewidth is determined after plotting
 
-        # Default color without a hue semantic should follow the color cycle
-        # Note, this is fairly complicated and awkward, I'd like a better way
-        # TODO and now with the ax business, this is just super annoying FIX!!
-        if "hue" not in self.variables:
-            if self.ax is None:
-                default_color = "C0" if color is None else color
-            else:
-                if fill:
-                    if self.var_types[self.data_variable] == "datetime":
-                        # Avoid drawing empty fill_between on date axis
-                        # https://github.com/matplotlib/matplotlib/issues/17586
-                        scout = None
-                        default_color = plot_kws.pop("facecolor", color)
-                        if default_color is None:
-                            default_color = "C0"
-                    else:
-                        artist = mpl.patches.Rectangle
-                        plot_kws = _normalize_kwargs(plot_kws, artist)
-                        scout = self.ax.fill_between([], [], color=color, **plot_kws)
-                        default_color = tuple(scout.get_facecolor().squeeze())
-                else:
-                    artist = mpl.lines.Line2D
-                    plot_kws = _normalize_kwargs(plot_kws, artist)
-                    scout, = self.ax.plot([], [], color=color, **plot_kws)
-                    default_color = scout.get_color()
[…]
