### 2021-01-30 ba086c2
Update README CI status badge to reflect GitHub Actions status
Closes #2453
