### 2021-01-30 999e175
Handle deprecation of ci=None in lineplot
Fixes #2441
