### 2021-01-30 ab4e6dc
Fix rugplot with datetime data
Fixes #2451
