### 2021-01-31 63e28ce
Handle deprecation of ci=None in lineplot (#2457)
Fixes #2441
