### 2021-01-31 6fd28be
Fix rugplot with datetime data (#2458)
Fixes #2451

(cherry picked from commit 136b462b30194f6579fd761d639118d0ceb66f55)
