### 2021-01-31 c4a7351
Set alpha properly on unfilled hist/kde plots
Fixes #2450
