### 2021-01-31 2236060
Generalize color comparison
