+++ seaborn/tests/test_distributions.py
@@ -35,6 +35,7 @@ from ..axisgrid import FacetGrid
 from .._testing import (
     assert_plots_equal,
     assert_legends_equal,
+    assert_colors_equal,
 )
 
 
@@ -123,16 +124,16 @@ class SharedAxesLevelTests:
 
         ax = plt.figure().subplots()
         self.func(data=long_df, x="y", ax=ax, **kwargs)
-        assert self.get_last_color(ax, **kwargs) == to_rgb("C0")
+        assert_colors_equal(self.get_last_color(ax, **kwargs), "C0", check_alpha=False)
 
         ax = plt.figure().subplots()
         self.func(data=long_df, x="y", ax=ax, **kwargs)
         self.func(data=long_df, x="y", ax=ax, **kwargs)
-        assert self.get_last_color(ax, **kwargs) == to_rgb("C1")
+        assert_colors_equal(self.get_last_color(ax, **kwargs), "C1", check_alpha=False)
 
         ax = plt.figure().subplots()
         self.func(data=long_df, x="y", color="C2", ax=ax, **kwargs)
-        assert self.get_last_color(ax, **kwargs) == to_rgb("C2")
+        assert_colors_equal(self.get_last_color(ax, **kwargs), "C2", check_alpha=False)
 
 
 class TestRugPlot(SharedAxesLevelTests):
@@ -141,7 +142,7 @@ class TestRugPlot(SharedAxesLevelTests):
 
     def get_last_color(self, ax, **kwargs):
 
-        return to_rgb(ax.collections[-1].get_color()[0])
+        return ax.collections[-1].get_color()
 
     def assert_rug_equal(self, a, b):
 
@@ -302,9 +303,9 @@ class TestKDEPlotUnivariate(SharedAxesLevelTests):
     def get_last_color(self, ax, fill=True):
 
         if fill:
-            return to_rgb(ax.collections[-1].get_facecolor()[0])
+            return ax.collections[-1].get_facecolor()
         else:
-            return to_rgb(ax.lines[-1].get_color())
+            return ax.lines[-1].get_color()
 
     @pytest.mark.parametrize("fill", [True, False])
     def test_color(self, long_df, fill):
@@ -315,11 +316,11 @@ class TestKDEPlotUnivariate(SharedAxesLevelTests):
 
             ax = plt.figure().subplots()
             self.func(data=long_df, x="y", facecolor="C3", fill=True, ax=ax)
-            assert self.get_last_color(ax) == to_rgb("C3")
+            assert_colors_equal(self.get_last_color(ax), "C3", check_alpha=False)
 
             ax = plt.figure().subplots()
             self.func(data=long_df, x="y", fc="C4", fill=True, ax=ax)
-            assert self.get_last_color(ax) == to_rgb("C4")
+            assert_colors_equal(self.get_last_color(ax), "C4", check_alpha=False)
 
     @pytest.mark.parametrize(
         "variable", ["x", "y"],
@@ -442,8 +443,8 @@ class TestKDEPlotUnivariate(SharedAxesLevelTests):
         palette = color_palette()
 
         for line, fill, color in zip(lines, fills, palette):
-            assert line.get_color() == color
-            assert tuple(fill.get_facecolor().squeeze()) == color + (.25,)
+            assert_colors_equal(line.get_color(), color)
+            assert_colors_equal(fill.get_facecolor(), to_rgba(color, .25))
 
     def test_hue_stacking(self, long_df):
 
@@ -516,33 +517,26 @@ class TestKDEPlotUnivariate(SharedAxesLevelTests):
     def test_color_cycle_interaction(self, flat_series):
 
         color = (.2, 1, .6)
-        C0, C1 = to_rgb("C0"), to_rgb("C1")
 
         f, ax = plt.subplots()
         kdeplot(flat_series)
[…]
