### 2021-01-31 548b854
Set alpha properly on unfilled hist/kde plots (#2460)
* Set alpha properly on unfilled hist/kde plots

Fixes #2450

* Fix typo in kdeplot docstring

* Generalize color comparison
