+++ seaborn/_testing.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib as mpl
+from matplotlib.colors import to_rgb, to_rgba
 from numpy.testing import assert_array_equal
 
 
@@ -69,3 +70,21 @@ def assert_plots_equal(ax1, ax2, labels=True):
     if labels:
         assert ax1.get_xlabel() == ax2.get_xlabel()
         assert ax1.get_ylabel() == ax2.get_ylabel()
+
+
+def assert_colors_equal(a, b, check_alpha=True):
+
+    def handle_array(x):
+
+        if isinstance(x, np.ndarray):
+            if x.ndim > 1:
+                x = np.unique(x, axis=0).squeeze()
+            if x.ndim > 1:
+                raise ValueError("Color arrays must be 1 dimensional")
+        return x
+
+    a = handle_array(a)
+    b = handle_array(b)
+
+    f = to_rgba if check_alpha else to_rgb
+    assert f(a) == f(b)
