### 2021-02-01 c56b0c4
Test bivariate normalization
