### 2021-02-01 d0acb8c
Add percent-based normalization in histplot (#2461)
* Add percent-based normalization in histplot

* Test bivariate normalization
