### 2021-02-01 23d2442
Add percent-based normalization in histplot (#2461)
* Add percent-based normalization in histplot

* Test bivariate normalization

(cherry picked from commit d0acb8cd8308d66cb95795e2fdad99c4220403d8)
