### 2021-02-01 dc30328
Fix multiple resolution when hue variable has no name
Fixes #2452

(cherry picked from commit 008f7e0e030681f8047c7d628e6bb3fc3d23e6ff)
