### 2021-02-01 dbceeaf
Don't try to resolve multiples if there aren't any
