### 2021-02-03 b9dbfd4
Simplify logic that ignored multiple when no hue was assigned
