### 2021-02-03 ab70c30
Fix typo and remove outdated comment
