### 2021-02-03 17c0f62
Fix multiple resolution when hue variable has no name (#2462)
* Fix multiple resolution when hue variable has no name

Fixes #2452

(cherry picked from commit 008f7e0e030681f8047c7d628e6bb3fc3d23e6ff)

* Remove unused kwarg

* Don't try to resolve multiples if there aren't any

* Simplify logic that ignored multiple when no hue was assigned

* Fix typo and remove outdated comment

(cherry picked from commit beb7f822895375143de396659cb562465ade25d7)
