### 2021-02-03 beb7f82
Fix multiple resolution when hue variable has no name (#2462)
* Fix multiple resolution when hue variable has no name

Fixes #2452

(cherry picked from commit 008f7e0e030681f8047c7d628e6bb3fc3d23e6ff)

* Remove unused kwarg

* Don't try to resolve multiples if there aren't any

* Simplify logic that ignored multiple when no hue was assigned

* Fix typo and remove outdated comment
