### 2021-02-05 7ad4623
Fix faceted bivariate histogram normalization at displot (#2468)
Fixes #2465

(cherry picked from commit 9066ad1e6c8daec64a95cb3a29222c1621735261)
