### 2021-02-05 9066ad1
Fix faceted bivariate histogram normalization at displot (#2468)
Fixes #2465
