### 2021-02-05 283a485
Don't fail when docstrings have been stripped by -OO mode
Fixes #2470
