+++ seaborn/_docstrings.py
@@ -33,8 +33,8 @@ class DocstringComponents:
                 # If Python is run with -OO, it will strip docstrings and our lookup
                 # from self.entries will fail. We check for __debug__, which is actually
                 # set to False by -O (it is True for normal execution).
-                # But we only really want to see an error when building the docs,  it's
-                # not an error so this slight inconsistency is fine.
+                # But we only want to see an error when building the docs;
+                # not something users should see, so this slight inconsistency is fine.
                 if __debug__:
                     raise err
                 else:
