### 2021-02-05 8f83a4d
Don't fail when docstrings have been stripped by -OO mode (#2473)
* Don't fail when docstrings have been stripped by -OO mode

Fixes #2470

* Fix comment

* Clarify language
