+++ seaborn/_docstrings.py
@@ -23,11 +23,22 @@ class DocstringComponents:
         self.entries = entries
 
     def __getattr__(self, attr):
-        """Provided dot access to entries."""
+        """Provide dot access to entries for clean raw docstrings."""
         if attr in self.entries:
             return self.entries[attr]
         else:
-            return self.__getattribute__(attr)
+            try:
+                return self.__getattribute__(attr)
+            except AttributeError as err:
+                # If Python is run with -OO, it will strip docstrings and our lookup
+                # from self.entries will fail. We check for __debug__, which is actually
+                # set to False by -O (it is True for normal execution).
+                # But we only want to see an error when building the docs;
+                # not something users should see, so this slight inconsistency is fine.
+                if __debug__:
+                    raise err
+                else:
+                    pass
 
     @classmethod
     def from_nested_components(cls, **kwargs):
