### 2021-02-08 #3021 @whedon
[REVIEW]: seaborn: statistical data visualization
**Submitting author:** @mwaskom (<a href="http://orcid.org/0000-0002-9817-6869">Michael Waskom</a>)
**Repository:** <a href="https://github.com/mwaskom/seaborn" target ="_blank">https://github.com/mwaskom/seaborn</a>
**Version:** v0.11.1

    - *Yes - the 3-Clause BSD License*
    - *Yes - mwaskom has 1,841 commits. There is a broad community that has contributed to seaborn, but it is appropriate to have mwaskom as the sole author.*
    - *seaborn allows for easy data visualization in python and has the added advantage of integrating closely with `matplotlib` and `pandas`*
    - *seaborn has been ongoing since 2012 with over 2700 commits at this point in time*
    - *according to GitHub, there are over 2300 packages which depend on seaborn*
    - *according to Zenodo, there are 196 papers that cite seaborn* https://zenodo.org/record/4379347#.YFI2JGRKhhE

    - *Yes, [here](https://seaborn.pydata.org/installing.html#)*
    - Test functional claims from the paper:
    - [x] opinionated semantic mappings
    - [x] acceptance of pandas and numpy
    - [x] themes
    - [x] allowing `matplotlib` "add-ons"
    - I have not found any explicit claims in the paper or in the documentation.

    - *This section is in the paper and it addresses the problems which are solved*
    - ***The target audience is not explicitly defined in this section***
    -  *Yes, [here](https://seaborn.pydata.org/installing.html#)*
    - *specific case examples are provided in the [documentation](https://seaborn.pydata.org/examples/index.html).*
    - *a full example of how seaborn can be used in a project is linked [here](https://github.com/WagnerLabPapers/Waskom_PNAS_2017) and at the end of the paper*
    - *Yes, [here](https://seaborn.pydata.org/api.html)*
    - *Yes, detailed in the [README](https://github.com/mwaskom/seaborn)*
    - *Yes, detailed in the [README](https://github.com/mwaskom/seaborn)*

    - *as stated above,* ***The target audience is not explicitly defined in this section***
    - *Yes, this is very clear in the paper*
