### 2021-02-10 6183f1e
Fix histplot shrink with non-discrete bins (#2477)
The code for shifting the shrunken bars assumed that discrete binning
was in effect. This is probably the only situation where shrinking
really makes sense, but there was no prevention or warning of getting
an innacurate result when using it with continuous bins.

Fixes #2476
