### 2021-02-10 8fb6629
Fix histplot shrink with non-discrete bins (#2477)
The code for shifting the shrunken bars assumed that discrete binning
was in effect. This is probably the only situation where shrinking
really makes sense, but there was no prevention or warning of getting
an innacurate result when using it with continuous bins.

Fixes #2476

(cherry picked from commit 6183f1e3b0fc2b56ef441acfe1f9f63a13ae4beb)
