### 2021-02-17 b528ba2
update matplotlib canonical... (#2482)
.. this will help with inter sphinx warnings.
