### 2021-02-19 74f460c
Replace version numbering interface
As the distutils.version is depecrating, it was replaced
with the packaging.version. In order to give support for all
version numbering, according to PEP 440, parse was used.
