+++ seaborn/tests/test_utils.py
@@ -18,7 +18,7 @@ from pandas.testing import (
     assert_frame_equal,
 )
 
-from distutils.version import LooseVersion
+from packaging import version
 
 from .. import utils, rcmod
 from ..utils import (
@@ -325,14 +325,14 @@ def test_locator_to_legend_entries():
     locator = mpl.ticker.LogLocator(numticks=5)
     limits = (5, 1425)
     levels, str_levels = utils.locator_to_legend_entries(locator, limits, int)
-    if LooseVersion(mpl.__version__) >= "3.1":
+    if version.parse(mpl.__version__) >= version.parse("3.1"):
         assert str_levels == ['10', '100', '1000']
 
     limits = (0.00003, 0.02)
     levels, str_levels = utils.locator_to_legend_entries(
         locator, limits, float
     )
-    if LooseVersion(mpl.__version__) >= "3.1":
+    if version.parse(mpl.__version__) >= version.parse("3.1"):
         assert str_levels == ['1e-04', '1e-03', '1e-02']
 
 
