+++ seaborn/tests/test_distributions.py
@@ -1,5 +1,5 @@
 import itertools
-from distutils.version import LooseVersion
+from packaging import version
 
 import numpy as np
 import matplotlib as mpl
@@ -1376,7 +1376,7 @@ class TestHistPlotUnivariate(SharedAxesLevelTests):
             histplot(long_df, x="s", discrete=True, element="poly")
 
     @pytest.mark.skipif(
-        LooseVersion(np.__version__) < "1.17",
+        version.parse(np.__version__) < version.parse("1.17"),
         reason="Histogram over datetime64 requires numpy >= 1.17",
     )
     def test_datetime_scale(self, long_df):
