+++ seaborn/tests/test_categorical.py
@@ -10,7 +10,7 @@ from matplotlib.colors import rgb2hex, to_rgb, to_rgba
 import pytest
 from pytest import approx
 import numpy.testing as npt
-from distutils.version import LooseVersion
+from packaging import version
 from numpy.testing import (
     assert_array_equal,
     assert_array_less,
@@ -1631,7 +1631,7 @@ class SharedScatterTests(SharedAxesLevelTests):
         self.func(data=long_df, x="a", y="y", facecolor="C4", ax=ax)
         assert self.get_last_color(ax) == to_rgba("C4")
 
-        if LooseVersion(mpl.__version__) >= "3.1.0":
+        if version.parse(mpl.__version__) >= version.parse("3.1.0"):
             # https://github.com/matplotlib/matplotlib/pull/12851
 
             ax = plt.figure().subplots()
@@ -1646,7 +1646,7 @@ class SharedScatterTests(SharedAxesLevelTests):
 
         keys = ["c", "facecolor", "facecolors"]
 
-        if LooseVersion(mpl.__version__) >= "3.1.0":
+        if version.parse(mpl.__version__) >= version.parse("3.1.0"):
             # https://github.com/matplotlib/matplotlib/pull/12851
             keys.append("fc")
 
@@ -2047,7 +2047,7 @@ class SharedScatterTests(SharedAxesLevelTests):
         # (Even though visual output is ok -- so it's not an actual bug).
         # I'm not exactly sure why, so this version check is approximate
         # and should be revisited on a version bump.
-        if LooseVersion(mpl.__version__) < "3.1":
+        if version.parse(mpl.__version__) < version.parse("3.1"):
             pytest.xfail()
 
         ax = plt.figure().subplots()
@@ -3279,7 +3279,7 @@ class TestBoxenPlotter(CategoricalFixture):
         plt.close("all")
 
     @pytest.mark.skipif(
-        LooseVersion(pd.__version__) < "1.2",
+        version.parse(pd.__version__) < version.parse("1.2"),
         reason="Test requires pandas>=1.2")
     def test_Float64_input(self):
         data = pd.DataFrame(
