### 2021-02-26 f263b59
Don't pass a set object to ListedColormap
Addresses the issue in https://github.com/matplotlib/matplotlib/issues/19544
