### 2021-02-26 090e006
Handle multiple LineCollection color kwargs
