### 2021-02-26 ff4070b
Test side colors size homogeneity test
