### 2021-03-06 ba4bd0f
Don't pass a set to ListedColormap in clustermap (#2490)
* Don't pass a set object to ListedColormap

Addresses the issue in https://github.com/matplotlib/matplotlib/issues/19544

* Handle multiple LineCollection color kwargs

* Test side colors size homogeneity test
