### 2021-03-06 49fbd35
Don't pass a set to ListedColormap in clustermap (#2490)
* Don't pass a set object to ListedColormap

Addresses the issue in https://github.com/matplotlib/matplotlib/issues/19544

* Handle multiple LineCollection color kwargs

* Test side colors size homogeneity test

(cherry picked from commit ba4bd0fa0a90b2bd00cb62c2b4a5e38013a73ac6)
