### 2021-03-06 e88bb64
Improve log_scale parameter documentation
Partly addresses #2454
