### 2021-03-12 476f89d
Fix log-scaled ecdf plots (fixes #2502)
