### 2021-03-12 e60fea9
Fix log-scaled rug plot
