### 2021-03-17 #3021 @Sara-ShiHo
Hi @whedon @lpantano is it possible to resend the invitation at https://github.com/openjournals/joss-reviews/invitations
