### 2021-03-23 #3021 @lpantano
@mwaskom, can you address this issue raised by @Sara-ShiHo:

```
A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

    OK -> This section is in the paper and it addresses the problems which are solved
    NEED TO ADD -> The target audience is not explicitly defined in this section
```

I think after that, it would be ready to be accepted. Thanks!
