### 2021-03-23 #3021 @mwaskom
@lpantano sure, sounds good. Also a couple of people who have seen the draft manuscript have suggested that it might make sense for a paper about a data visualization library to have an example data visualization. I guess that's reasonable. Would it make sense to add one at this stage?
