### 2021-03-23 7859883
Add example figure to the paper
