### 2021-03-23 #3021 @mwaskom
OK I added two commits:

https://github.com/mwaskom/seaborn/commit/2be812a0d5f3c7ef950507c505baa034137ab990 addresses the target audience. seaborn is domain-general so the target audience is pretty wide; I think it's most helpful to be explicit about some kinds of visualizations that are explicitly out of scope.

https://github.com/mwaskom/seaborn/commit/43763350eb8138c9eca4d907c6188a734f08f5cc adds an example script and figure. This seems useful, but I'm happy to revert it if you think it's too late in the process to add it.
