### 2021-03-24 #3021 @Sara-ShiHo
@lpantano & @mwaskom I reread the updated paper and I updated my checklist. One last thing (seeing the example section reminded me) - do you think it is necessary to specify seaborn's current version number in the example section or anywhere else?
