### 2021-03-24 #3021 @mwaskom
Happy to do whatever is common for JOSS papers.
