### 2021-03-24 6d84f82
Fix log scaling in distribution plots (#2504)
* Improve log_scale parameter documentation

Partly addresses #2454

* Fix log-scaled ecdf plots (fixes #2502)

* Fix log-scaled rug plot
