### 2021-03-24 b5ccad9
Fix log scaling in distribution plots (#2504)
* Improve log_scale parameter documentation

Partly addresses #2454

* Fix log-scaled ecdf plots (fixes #2502)

* Fix log-scaled rug plot

(cherry picked from commit 6d84f82b8edeb00963d74fa5e29b5c9fc3e7248c)
