+++ doc/releases/v0.12.0.txt
@@ -40,3 +40,5 @@ v0.12.0 (Unreleased)
 - Following `NEP29 <https://numpy.org/neps/nep-0029-deprecation_policy.html>`_, dropped support for Python 3.6 and bumped the minimally-supported versions of the library dependencies.
 
 - |Fix| In :func:`histplot` and :func:`kdeplot`, fixed a bug where the `alpha` parameter was ignored when `fill=False` (:pr:`2460`).
+
+- |Fix| In :func:`ecdfplot` and :func:`rugplot`, fixed a bug where results were incorrect with a log-scaled data axis (:pr:`2504`).
