### 2021-03-24 4ea21a7
Fix histplot auto linewidth on log scaled axis
Fixes #2513
