### 2021-03-25 e54c2d7
Fix auto linewidth for categorical histograms on the y axis
Fixes #2523
