### 2021-03-25 9776fb8
Fix histplot auto line width with log scale or categorical y axis (#2522)
* Fix histplot auto linewidth on log scaled axis

Fixes #2513

* Fix auto linewidth for categorical histograms on the y axis

Fixes #2523

(cherry picked from commit 0f538fb40ad04174a9298f9403783308f9bbb370)
