### 2021-03-26 #3021 @lpantano
It is a good suggestion, @mwaskom, would you mind to add the version used on the paper under the example section? Not all the papers show code, so there is no mandatory rule, and it should hurt to  have it. Thanks @Sara-ShiHo for the sugestion.
