### 2021-03-29 #3021 @mwaskom
OK I have made a slight mess of things by naming the tag with the same string as the branch containing the paper.

In any case, I think it should be resolved.

Provided that the tagged release **does not** need to complain the paper source, this checklist should be completed. @lpantano please let me know if I need to change that.

Release was made with version tag [`joss_paper`](https://github.com/mwaskom/seaborn/releases/tag/joss_paper), which points at the same commit as `v0.11.1`.

The DOI for the release is [10.5281/zenodo.4645478](https://zenodo.org/record/4645478#.YGITw0hKhTY).
