+++ .readthedocs.yaml
@@ -1,0 +1,9 @@
+version: 2
+python:
+  version: 3.8
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - all
+    - requirements: doc/requirements.txt
