### 2021-03-31 #3021 @mwaskom
Hm I thought I saw in another issue that it's possible to direct editorial tasks to specific branches, but I'm not sure.
