### 2021-04-02 b9c7fae
x_estimator bars now inherit scatter_kws alpha
x_estimator error bars were previously always opaque, but now inherit alpha settings from scatterplot settings, since the error bars replace the scatterplot
