+++ paper.bib
@@ -1,7 +1,7 @@
 
 @article{Cumming:2005,
 author = {Cumming, Geoff and Finch, Sue},
-title = {{Inference by eye: confidence intervals and how to read pictures of data.}},
+title = {{Inference by eye: confidence intervals and how to read pictures of data}},
 journal = {The American Psychologist},
 year = {2005},
 volume = {60},
@@ -51,7 +51,7 @@ doi = {10.1037/0003-066X.60.2.170},
 @InProceedings{ McKinney:2010,
   author    = { Wes McKinney },
   title     = { Data Structures for Statistical Computing in Python },
-  booktitle = { Proceedings of the 9th Python in Science Conference },
+  booktitle = {{ Proceedings of the 9th Python in Science Conference }},
   pages     = { 51 - 56 },
   year      = { 2010 },
   editor    = { St\'efan van der Walt and Jarrod Millman },
