### 2021-04-05 #3021 @danielskatz
👋 @mwaskom - there are two minor changes in https://github.com/mwaskom/seaborn/pull/2542, one of which might not matter...  If you can merge this, I can go ahead and publish the paper
