### 2021-04-06 #3021 @arfon
@mwaskom - apologies for the fussing with your BibTeX file but could you also merge this PR? https://github.com/mwaskom/seaborn/pull/2544

I think the BibTeX library we're using has a bug with whitespace.
