+++ paper.bib
@@ -50,8 +50,8 @@ doi = {10.1037/0003-066X.60.2.170},
 
 @InProceedings{ McKinney:2010,
   author    = { Wes McKinney },
-  title     = { Data Structures for Statistical Computing in Python },
-  booktitle = {{ Proceedings of the 9th Python in Science Conference }},
+  title     = {Data Structures for Statistical Computing in Python},
+  booktitle = {{Proceedings of the 9th Python in Science Conference}},
   pages     = { 51 - 56 },
   year      = { 2010 },
   editor    = { St\'efan van der Walt and Jarrod Millman },
